Bug fix: print `attrs` length and count from 1
Had an error where the length of the object in the loop was printed
instead of the list the object was part of. This will now print the
length of the list.

Also, the iteration index was printed as is. Since it starts at 0, it
would end at `len(list) - 1`. Now, index + 1 is printed.